+++ app/settings/page.tsx
@@ -1,69 +1,30 @@
 'use client';
 
 import { Sidebar } from '@/components/sidebar';
-import {
-  AlertDialog,
-  AlertDialogAction,
-  AlertDialogCancel,
-  AlertDialogContent,
-  AlertDialogDescription,
-  AlertDialogFooter,
-  AlertDialogHeader,
-  AlertDialogTitle,
-  AlertDialogTrigger,
-} from '@/components/ui/alert-dialog';
 import { Button } from '@/components/ui/button';
 import { Switch } from '@/components/ui/switch';
-import { useAuth } from '@/lib/auth-context';
-import { authMfaService } from '@/services/auth-mfa.service';
 import {
-  AlertTriangle,
   ArrowLeft,
   AudioLines,
   Brain,
-  Loader2,
   MessageSquare,
   Moon,
-  Shield,
-  Smartphone,
   Sun,
-  Trash2,
 } from 'lucide-react';
 import { useTheme } from 'next-themes';
 import Link from 'next/link';
-import { Suspense, useCallback, useEffect, useState } from 'react';
-import { toast } from 'sonner';
+import { Suspense, useEffect, useState } from 'react';
 import { useSettings } from '../../lib/settings-context';
 
 export default function SettingsPage() {
   const { theme, setTheme, systemTheme } = useTheme();
   const { settings, updateSettings } = useSettings();
-  const { tokens } = useAuth();
 
   const [mounted, setMounted] = useState(false);
-  const [mfaFactors, setMfaFactors] = useState<any[]>([]);
-  const [isLoadingMfa, setIsLoadingMfa] = useState(false);
-  const [isActionLoading, setIsActionLoading] = useState(false);
-
-  const loadMfaStatus = useCallback(async () => {
-    if (!tokens?.access_token) return;
-    setIsLoadingMfa(true);
-    try {
-      const factors = await authMfaService.listFactors(tokens.access_token);
-      setMfaFactors(factors.active || []);
-    } catch (error) {
-      console.error('Failed to load MFA status:', error);
-    } finally {
-      setIsLoadingMfa(false);
-    }
-  }, [tokens]);
 
   useEffect(() => {
     setMounted(true);
-    if (tokens?.access_token) {
-      loadMfaStatus();
-    }
-  }, [tokens, loadMfaStatus]);
+  }, []);
 
   if (!mounted) return null;
 
@@ -82,22 +43,6 @@ export default function SettingsPage() {
 
   const handleToggleVoiceSuggestions = (checked: boolean) => {
     updateSettings({ showVoiceSuggestions: checked });
-  };
-
-  const handleUnenrollMFA = async (factorId: string) => {
-    if (!tokens?.access_token) return;
-    setIsActionLoading(true);
-    try {
-      await authMfaService.unenrollFactor(factorId, tokens.access_token);
-      toast.success('Đã hủy liên kết thiết bị xác thực');
-      await loadMfaStatus();
-    } catch (error: any) {
-      toast.error(
-        'Không thể hủy liên kết: ' + (error.message || 'Lỗi không xác định')
-      );
-    } finally {
-      setIsActionLoading(false);
-    }
   };
 
   const currentTheme = theme === 'system' ? systemTheme : theme;
@@ -251,111 +196,6 @@ export default function SettingsPage() {
                   checked={settings.showVoiceSuggestions}
                   onCheckedChange={handleToggleVoiceSuggestions}
                 />
-              </div>
-            </div>
-
-            {/* Quản lý MFA */}
-            <div className="bg-slate-50 dark:bg-slate-900 rounded-[2rem] p-8 mb-8 border border-slate-200 dark:border-slate-800 shadow-sm overflow-hidden relative">
-              <div className="absolute top-0 right-0 p-8 opacity-[0.03] dark:opacity-[0.05] pointer-events-none">
-                <Shield size={120} />
-              </div>
-
-              <div className="flex items-center gap-4 mb-8">
-                <div className="p-3 bg-indigo-100 dark:bg-indigo-900/40 text-indigo-600 dark:text-indigo-400 rounded-2xl">
-                  <Shield size={28} />
-                </div>
-                <div>
-                  <h2 className="text-xl font-bold text-slate-900 dark:text-white">
-                    Bảo mật 2 bước (MFA)
-                  </h2>
-                  <p className="text-sm text-slate-600 dark:text-slate-400">
-                    Bảo vệ tài khoản của bạn bằng mã xác thực OTP
-                  </p>
-                </div>
-              </div>
-
-              <div className="space-y-4 relative z-10">
-                {isLoadingMfa ? (
-                  <div className="flex items-center justify-center py-8">
-                    <Loader2 className="w-8 h-8 animate-spin text-indigo-500" />
-                  </div>
-                ) : mfaFactors.length > 0 ? (
-                  <div className="space-y-4">
-                    {mfaFactors.map((factor) => (
-                      <div
-                        key={factor.id}
-                        className="flex items-center justify-between p-5 bg-white dark:bg-slate-800/50 rounded-2xl border border-slate-200 dark:border-slate-800 group transition-all hover:shadow-md"
-                      >
-                        <div className="flex items-center gap-4">
-                          <div className="w-12 h-12 flex items-center justify-center bg-indigo-50 dark:bg-indigo-900/20 text-indigo-600 dark:text-indigo-400 rounded-xl">
-                            <Smartphone size={24} />
-                          </div>
-                          <div>
-                            <p className="font-bold text-slate-900 dark:text-white">
-                              {factor.friendly_name || 'Thiết bị xác thực'}
-                            </p>
-                            <p className="text-xs text-slate-500 flex items-center gap-1">
-                              <span className="w-2 h-2 rounded-full bg-green-500"></span>
-                              Đang hoạt động (TOTP)
-                            </p>
-                          </div>
-                        </div>
-
-                        <AlertDialog>
-                          <AlertDialogTrigger asChild>
-                            <Button
-                              variant="ghost"
-                              size="icon"
-                              className="text-slate-400 hover:text-red-500 hover:bg-red-50 dark:hover:bg-red-950/30 rounded-xl transition-colors"
-                              disabled={isActionLoading}
-                            >
-                              <Trash2 size={18} />
-                            </Button>
-                          </AlertDialogTrigger>
-                          <AlertDialogContent className="bg-white dark:bg-slate-900 border-slate-200 dark:border-slate-800 rounded-3xl p-8">
-                            <AlertDialogHeader>
-                              <div className="w-16 h-16 bg-red-100 dark:bg-red-900/20 text-red-600 dark:text-red-400 rounded-2xl flex items-center justify-center mb-4">
-                                <AlertTriangle size={32} />
-                              </div>
-                              <AlertDialogTitle className="text-2xl font-bold text-slate-900 dark:text-white">
-                                Hủy liên kết MFA?
-                              </AlertDialogTitle>
-                              <AlertDialogDescription className="text-slate-600 dark:text-slate-400 text-base leading-relaxed">
-                                Hành động này sẽ tắt bảo mật 2 bước cho tài
-                                khoản của bạn. Bạn sẽ cần phải đăng ký lại nếu
-                                muốn bật lại tính năng này.
-                              </AlertDialogDescription>
-                            </AlertDialogHeader>
-                            <AlertDialogFooter className="mt-8 gap-3">
-                              <AlertDialogCancel className="h-12 px-6 rounded-xl border-slate-200 dark:border-slate-800 text-slate-600 dark:text-slate-400 font-bold hover:bg-slate-100 dark:hover:bg-slate-800 transition-all">
-                                Quay lại
-                              </AlertDialogCancel>
-                              <AlertDialogAction
-                                onClick={() => handleUnenrollMFA(factor.id)}
-                                className="h-12 px-6 rounded-xl bg-red-600 hover:bg-red-500 text-white font-bold shadow-lg shadow-red-600/20 transition-all border-none"
-                              >
-                                {isActionLoading
-                                  ? 'Đang xử lý...'
-                                  : 'Xác nhận Hủy'}
-                              </AlertDialogAction>
-                            </AlertDialogFooter>
-                          </AlertDialogContent>
-                        </AlertDialog>
-                      </div>
-                    ))}
-                  </div>
-                ) : (
-                  <div className="text-center py-10 bg-white dark:bg-slate-800/30 rounded-3xl border border-dashed border-slate-300 dark:border-slate-700">
-                    <p className="text-slate-500 dark:text-slate-400 mb-6">
-                      Bạn chưa kích hoạt bảo mật 2 bước.
-                    </p>
-                    <Link href="/auth/mfa">
-                      <Button className="bg-indigo-600 hover:bg-indigo-500 text-white rounded-xl font-bold px-8 h-12 shadow-lg shadow-indigo-600/20 transition-all">
-                        Kích hoạt ngay
-                      </Button>
-                    </Link>
-                  </div>
-                )}
               </div>
             </div>
           </div>
